Initial working version.
This is a bit primitive in that it can't work out the NetROM callsign
for the port by itself, so bind fails if you don't set it specifically.
It also doesn't seem to be able to look up NetROM aliases, so you
have to find the callsign associated with the alias and use that as
the destination.